split unit tests into several files for maintainability
